Make getBlocksByName private
